+++ src/components/common/NavigationBar/Header/components/AuthButtons.tsx
