+++ src/components/common/NavigationBar/Header/components/SearchBar.tsx
